separated character detail elements. created landscape and portrait views.
